requirements.txt
# Backend
fastapi
uvicorn
python-dotenv
requests
azure-storage-blob
sentence-transformers
faiss-cpu

# Frontend
streamlit
